pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hierarchical-demand-reconciliation-with-temporal-anomaly-feedback"
version = "1.0.0"
description = "Closed-loop hierarchical forecasting system with temporal anomaly feedback"
authors = [{name = "Alireza Shojaei"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "torch>=2.0.0",
    "pytorch-lightning>=2.0.0",
    "pytorch-forecasting>=0.10.0",
    "darts>=0.25.0",
    "statsmodels>=0.14.0",
    "sktime>=0.24.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    "mlflow>=2.7.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "pyyaml>=6.0.0",
    "tqdm>=4.65.0",
    "kaggle>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
jupyter = [
    "jupyterlab>=4.0.0",
    "ipywidgets>=8.0.0",
]

[project.urls]
Homepage = "https://github.com/alireza-shojaei/hierarchical-demand-reconciliation-with-temporal-anomaly-feedback"
Repository = "https://github.com/alireza-shojaei/hierarchical-demand-reconciliation-with-temporal-anomaly-feedback"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing"